PPC64: beginnings of a SEC/IPL
This is totally harder than it looks... It took me a while to figure out that v2 ABI is completely
unusable in this PE/COFF environment because you end up with a ton of relocations like:
000000001000  000f00000029 R_PPC64_ADDR16_HI 000000000000a008 .TOC. + 0
000000001004  000f00000027 R_PPC64_ADDR16_HI 000000000000a008 .TOC. + 0
00000000100c  000f0000006e R_PPC64_ADDR16_HI 000000000000a008 .TOC. + 0
000000001010  000f00000004 R_PPC64_ADDR16_LO 000000000000a008 .TOC. + 0

...which are impossible to describe in base relocs. Fortunately
we still have the v1 ABI, which although is used for BE PPC64,
we can use just fine for LE as well.  We have to deal with
unfortunate bullshit like the function descriptors and awkward
ABIv2 interaction (since I'm launching the ipl from a modified
copy of my pp64le_hello example), but the relocations become
trivial:
000000001000  000500000026 R_PPC64_ADDR64    0000000000001018 ._ModuleEntryPoint + 0
000000001008  000000000033 R_PPC64_TOC                          0

It also took me a while to figure out that I should fix up R_PPC64_TOC relocs
as well >:-{.

This gets called like so:
{
          uint64_t *desc;
          desc = (uint64_t *) (*(uint64_t *) 0x40000000);
          void (*fn)(void (*ps)(char *), void *opal_base) = (void *) *desc;
          printk("Calling 0x%x\n", fn);
          fn(puts, opal_base);
}

The FD is loaded at 0x40000000 by skiboot, and offset 0 contains a pointer
to a function descriptor (sigh). At some point I'll fix it to be just
a normal branch. The ModuleEntryPoint code is smart enough to load
its TOC pointer, at least, but there's a lot left to do...

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>